# docs/swagger.yaml
definitions:
  request.BusinessAccountCreateRequest:
    properties:
      description:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - description
    - name
    type: object
  request.JwtCreateRequest:
    properties:
      userID:
        type: string
    type: object
  request.UserCreateRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      firstName:
        minLength: 2
        type: string
      lastName:
        type: string
      password:
        maxLength: 16
        minLength: 8
        type: string
    required:
    - age
    - email
    - firstName
    - lastName
    - password
    type: object
  request.UserSignInRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 16
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  response.BusinessAccountResponse:
    properties:
      _id:
        type: string
      createdAt:
        type: string
      description:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  response.JwtResponse:
    properties:
      _id:
        type: string
      accessToken:
        type: string
      createdAt:
        type: string
      refreshToken:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  response.UserResponse:
    properties:
      _id:
        type: string
      age:
        type: integer
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact:
    email: alpha@gmail.com
    name: Alpha
  description: This is a sample swagger for alpha rest api
  title: Alpha Fiber Rest Api
  version: "1.0"
paths:
  /api/v1/alpha/business-account:
    get:
      consumes:
      - application/json
      description: get all business accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.BusinessAccountResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for get all business accounts
      tags:
      - Business Accounts
    post:
      consumes:
      - application/json
      description: saving new business account
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        schema:
          $ref: '#/definitions/request.BusinessAccountCreateRequest'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for saving new business account
      tags:
      - Business Accounts
  /api/v1/alpha/jwt:
    get:
      consumes:
      - application/json
      description: get all jwts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.JwtResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for get all jwts
      tags:
      - JWT
    post:
      consumes:
      - application/json
      description: saving new jwt
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        schema:
          $ref: '#/definitions/request.JwtCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for saving new jwt
      tags:
      - JWT
  /api/v1/alpha/jwt/refresh:
    post:
      consumes:
      - application/json
      description: saving new jwt
      parameters:
      - description: '{token}'
        in: header
        name: x-refresh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for saving new jwt
      tags:
      - JWT
  /api/v1/alpha/user:
    get:
      consumes:
      - application/json
      description: get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.UserResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for get all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: saving new user
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        schema:
          $ref: '#/definitions/request.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for saving new user
      tags:
      - User
  /api/v1/alpha/user/{userId}:
    get:
      consumes:
      - application/json
      description: get user by id
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method get user by given id
      tags:
      - User
  /api/v1/alpha/user/sign-in:
    post:
      consumes:
      - application/json
      description: sign up for user
      parameters:
      - description: Handle Request Body
        in: body
        name: requestBody
        schema:
          $ref: '#/definitions/request.UserSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: This method used for sign up to user
      tags:
      - User
swagger: "2.0"